Add files via upload
The Percolation.java implements the Weighted Quick Union algorithm into a grid to check whether the top and bottom are connected and know whether a system will percolate, the WeightedQuickUnionUF is to be used as a package as it contains the algorithm to run the Percolation algorithm. Weighted Quick Union is a very time efficient algorithm of time complexity O(log* n) to perform union operations between a given set of n integers.